change the structure, readme and add 2 trivias
the structure of the repository changed with trivias now stored into
sub-directories inside the trivias directory. The change was made to
be more readable when adding new trivia isntead of a giant file.
Another change was to split the question and answer into two separate
files so we can link to the question file on social medias.
Finally the README have been updated to reflect those change and
improve the contributing guidelines.